memory: fast path for local free
Since most free calls would be frees for local CPU we wish to quickly
find out if this is the case;

For that, we'll keep a CPU local cpu_id_mask and a CPU local expected cpu_id
value. We could then quickly mask the given pointer with the mask, and
see if it matches the expected cpu_id.

For foreign pointers, the expected cpu_id value would be all bits set,
and the condition would be impossible to satisfy on foreign thread.

Signed-off-by: Elazar Leibovich <elazarl@gmail.com>